refactor:いろいろやばい書き方をしている部分を直した。pathpidaやapi処理など
