+++ package.json
@@ -7,7 +7,7 @@
 		"build": "next build",
 		"start": "next start",
 		"lint": "biome check . && eslint .",
-		"lint:fix": "biome check . --write --unsafe && eslint . --fix",
+		"lint:fix": "biome check . --write --unsafe",
 		"format": "biome format ./src/**/* --write",
 		"check": "biome check --apply ./src/**/*",
 		"prepare": "husky install",
@@ -36,7 +36,8 @@
 		"react-dom": "^19.0.0",
 		"sonner": "^2.0.3",
 		"tailwind-merge": "^3.2.0",
-		"uuid": "^11.1.0"
+		"uuid": "^11.1.0",
+		"vaul": "^1.1.2"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "1.9.4",
