+++ src/app/page.tsx
@@ -1,11 +1,11 @@
-import { PostCard } from "@/components/post-card";
+import { PostCard } from "@/components/postCard";
 import { MainLayout } from "@/layouts/main/layout";
 import type { Post } from "@/types/post";
 import { v4 as uuidv4 } from "uuid";
 import { NewPost } from "./posts/newPost";
 
 async function fetchPosts(): Promise<Post[]> {
-	const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/posts`);
+	const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/v1/posts`);
 	if (!res.ok) throw new Error(`HTTP Error! status: ${res.status}`);
 	return await res.json();
 }
