+++ src/app/posts/newPost.tsx
@@ -20,6 +20,7 @@ import {
 } from "@/components/ui/select";
 import { breakPoints } from "@/constants/breakpoints";
 import { visibilityOptions } from "@/constants/visibilityOptions";
+import { createPost } from "@/lib/api/post";
 import { useMediaQuery } from "@mui/material";
 import {
 	CircleEllipsis,
@@ -52,25 +53,7 @@ export const NewPost = () => {
 
 	const handleSubmit = async () => {
 		const postData = async () => {
-			const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/posts`, {
-				method: "POST",
-				headers: {
-					"Content-Type": "application/json",
-				},
-				body: JSON.stringify({
-					post: {
-						content,
-						user_id: user?.id,
-					},
-				}),
-			});
-
-			if (!res.ok) {
-				const data = await res.json();
-				throw new Error(data.error || "投稿に失敗しました");
-			}
-
-			return res;
+			await createPost(content, user?.id);
 		};
 
 		toast.promise(postData(), {
