+++ src/layouts/main/layout.tsx
@@ -5,7 +5,7 @@ export const MainLayout = ({ children }: { children: React.ReactNode }) => {
 	return (
 		<div className="flex min-h-screen flex-col">
 			<Header />
-			<main className="mx-auto flex w-full max-w-3xl flex-grow flex-col items-center p-8">
+			<main className="mx-auto flex w-full max-w-3xl flex-grow flex-col items-center">
 				{children}
 			</main>
 			<Toaster />
